Change update.py to match chief-james's expecations.
chief-james expects update.py to unconditionally check out the ref that
is passed. Previously, update_code would treat the parameter passed to
it as a branch name to check out and update. This changes update_code to
treat the parameter passed as a ref to get from the server, matching
chief-james's expectations.

fixes bug 1197960